[android] User changeTime to avoid race conditions

+++ app-android/app/src/main/java/de/tutao/tutanota/alarms/EndType.java
@@ -4,11 +4,11 @@ public enum EndType {
 	NEVER, COUNT, UNTIL;
 
 
-	public static EndType get(int value) {
+	public static EndType get(long value) {
 		if (value > EndType.values().length - 1) {
 			throw new IllegalArgumentException("Unknown EndType: " + value);
 		}
-		return EndType.values()[value];
+		return EndType.values()[(int) value];
 	}
 }
