+++ app-android/app/src/main/java/de/tutao/tutanota/alarms/RepeatPeriod.java
@@ -3,10 +3,10 @@ package de.tutao.tutanota.alarms;
 public enum RepeatPeriod {
 	DAILY, WEEKLY, MONTHLY, ANNUALLY;
 
-	public static RepeatPeriod get(int value) {
+	public static RepeatPeriod get(long value) {
 		if (value > RepeatPeriod.values().length - 1) {
 			throw new IllegalArgumentException("Unknown RepeatPeriod: " + value);
 		}
-		return RepeatPeriod.values()[value];
+		return RepeatPeriod.values()[(int) value];
 	}
 }
